requirements.txt
requests
pyyaml
-e .
